Feat: like update bug
