+++ src/components/pages/timelinePage/Posts.js
@@ -1,24 +1,9 @@
 import { useState, useEffect } from "react";
 import Error from "../../ErrorComponent";
 import Post from "../../PostComponent";
-import { getAllRecentPosts } from "../../../services/linkr";
 
-export default function Posts({ posts, setPosts }) {
-	const [disabled, setDisabled] = useState(false);
-	const token = localStorage.getItem("linkrUserToken");
 
-	useEffect(() => {
-		getAllRecentPosts(token)
-			.then((resp) => {
-				const postsData = resp.data;
-				setPosts(postsData);
-			})
-			.catch((err) => {
-				console.error(err);
-				setDisabled(true);
-			});
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, []);
+export default function Posts({ posts, setPosts, disabled }) {
 
 	return (
 		<>
