+++ src/components/pages/timelinePage/index.js
@@ -1,12 +1,29 @@
+import { useState, useEffect } from 'react';
 import { TimelinePage, PageStyle } from "./style";
 import Posts from "./Posts";
 import TrendingComponent from "../../TrendingComponent";
 import HeaderComponent from "../../headerComponent";
 import Publish from "../../PublishComponent";
-import { useState } from 'react';
+import { getAllRecentPosts } from "../../../services/linkr";
 
 export default function Timeline() {
 	const [posts, setPosts] = useState([]);
+	const [disabled, setDisabled] = useState(false);
+	const token = localStorage.getItem("linkrUserToken");
+
+	useEffect(() => {
+		getAllRecentPosts(token)
+			.then((resp) => {
+				const postsData = resp.data;
+				setPosts(postsData);
+			})
+			.catch((err) => {
+				console.error(err);
+				setDisabled(true);
+			});
+		// eslint-disable-next-line react-hooks/exhaustive-deps
+	}, []);
+
 	return (
 		<>
 		<HeaderComponent />
@@ -16,7 +33,7 @@ export default function Timeline() {
 					<h2>timeline</h2>
 				</div>
 				<Publish setPosts={setPosts} />
-				<Posts posts={posts} setPosts={setPosts} />
+				<Posts posts={posts} setPosts={setPosts} disabled={disabled} />
 			</TimelinePage>
 			<TrendingComponent />
 		</PageStyle>
